+++ xenon-common/src/main/java/com/vmware/xenon/common/Service.java
@@ -219,6 +219,11 @@ public interface Service extends ServiceRequestSender {
          */
         TRANSACTION_PENDING,
 
+        /**
+         * Set by runtime. Service is stateless.
+         */
+        STATELESS,
+
         NONE
     }