+++ xenon-common/src/main/java/com/vmware/xenon/common/Utils.java
@@ -597,6 +597,10 @@ public class Utils {
             break;
         case TRANSACTION_PENDING:
             break;
+        case STATELESS:
+            antiReqs = EnumSet.of(ServiceOption.PERSISTENCE, ServiceOption.REPLICATION,
+                    ServiceOption.OWNER_SELECTION, ServiceOption.STRICT_UPDATE_CHECKING);
+            break;
         default:
             break;
         }