+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.1.1-SNAPSHOT
 
+* Validate service options for Stateless services.
+  Using PERSISTENCE, REPLICATION, OWNER_SELECTION, STRICT_UPDATE_CHECKING
+  will fail service start.
+
 ## 1.1.0
 
 * Remove infrastructure Utils.toDocumentBytes/fromDocumentBytes and related methods,